+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.Abstractions/Commands/Command.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace Allegro.Extensions.Cqrs.Abstractions.Commands;
+
+/// <summary>
+/// Cqrs base representation of command.
+/// </summary>
+public abstract record Command
+{
+    /// <summary>
+    /// Command identifier
+    /// </summary>
+    public string Id { get; } = Guid.NewGuid().ToString();
+}