+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.Abstractions/Commands/ICommandValidator.cs
@@ -6,7 +6,7 @@ namespace Allegro.Extensions.Cqrs.Abstractions.Commands;
 /// Marker interface that allows to execute more complicated validations before action execution.
 /// </summary>
 /// <typeparam name="T">Supported command type</typeparam>
-public interface ICommandValidator<T> where T : ICommand
+public interface ICommandValidator<T> where T : Command
 {
     /// <summary>
     /// Validates command before execution. Should throw ValidationException or other exception.