+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.Abstractions/Queries/IQueryHandler.cs
@@ -8,7 +8,7 @@ namespace Allegro.Extensions.Cqrs.Abstractions.Queries;
 /// </summary>
 /// <typeparam name="TQuery">Type of supported query</typeparam>
 /// <typeparam name="TResult">Type of data returned by query</typeparam>
-public interface IQueryHandler<in TQuery, TResult> where TQuery : IQuery<TResult>
+public interface IQueryHandler<in TQuery, TResult> where TQuery : Query<TResult>
 {
     /// <summary>
     /// Handles query execution. In most cases reading data from read-model directly.